+++ packages/DBTests/src/package-test.ts
@@ -3,7 +3,6 @@ import * as grok from 'datagrok-api/grok';
 
 import {runTests, tests, TestContext, test as _test, category, initAutoTests as initTests } from '@datagrok-libraries/test/src/test';
 import './connections/queries-test';
-import './sync/data-sync-test';
 import './benchmarks/benchmark';
 import './cache/cache-test';
 import './connections/table-query-test';